Implement GPMO for arbitrary polarization vectors
A new function, GPMO_ArbVec, has been added to
permanent_magnet_optimization.cpp. It is based closely on GPMO_baseline,
but requires the user to input an array of allowable unit vectors
(argument `pol_vectors`) for each dipole in the arrangement. The
greedy algorithm then iterates through each unit vector of each dipole
to find the one that yields the greates reduction in the R2 parameter.
In contrast to GPMO_baseline and the other GPMO implementations, this
entails a nested iteration through the components of the arbitrary
vectors.

In principle, GPMO_ArbVec can be run with `A_obj` and `b_obj` matrices
in any basis (Cartesian, cylindrical, toroidal), provided the allowable
polarization unit vectors in `pol_vectors` are specified in the same
basis.

In addition, the Python modules have been updates to enable calls to
GPMO_ArbVec. Users must enter the array of arbitrary unit vectors as
an the optional parameter `pol_vectors` when initializing an instance
of the PermanentMagnetGrid class, and set `algorithm` to `'ArbVec'`
when calling GPMO().

In this commit, GPMO_ArbVec has been shown to produce identical
solutions to GPMO_baseline if the arbitrary vectors for each dipole are
restricted to be the three Cartesian unit vectors or the three
cylindrical unit vectors.