Modify GPMOb procedure for arbitrary vectors
In the GPMO backtracking procedure for arbitrary polarization vectors
(GPMO_ArbVec_backtracking), a change is made in how nearby magnets
are selected for elimination if they are found to be have too large
of an angular difference between their polarization vectors.

For each magnet placed in array, the procedure iterates through
the nearest `Nadjacent` magnets and checks the angular difference
between the dipole moment of the adjacent magnet and the reference
magnet. Previously, the nearest of the adjacent magnets found to have
an angular difference exceeding `thresh_angle` would be eliminated
from the solution along with the reference magnet. In the present
version, the adjacent magnet with the greatest angular difference from
the reference magnet is removed along with the reference magnet,
provided the difference exceeds `thresh_angle`.