Add option to retain Ic=0 dipoles in FocusData
Previously, when importing dipole data from a .focus file, the
initializer of the FocusData class would exclude data on dipoles
from the file for which the `Ic` parameter was zero. However, there
are some use cases that would require all dipoles to be imported
irrespective of the value of `Ic`. For example, in an optimization
with multiple stages, the user may wish to keep certain dipoles fixed
during some stages but not others.

To accommodate these cases, the initialization routine now has a new
optional argument `keep_Ic_zeros`. If true, all dipoles will be
imported from the file; if false, only dipoles with Ic==1 will be
imported. The argument is false by default, so any existing code
that uses the FocusData class should not be affected.