Modify backtracking output message
In GPMO_ArbVec_backtracking, the output message printed after each
backtracking procedure is modified to not include "out of
`backtracking` possible dipoles". The reason for this change
is that the message is not correct for the initial backtracking
procedure performed on the initialized solution, in which all magnets
(which may be greater than or less than `backtracking`) are eligible
for removal.