Implement basic simsoptpp WireframeField class
WireframeField is introduced in the simsoptpp C++ coses as a subclass
of the MagneticField class. The class is closely modeled on the
BiotSavart class for computing magnetic fields from lists of Coil
class instances.

The key method of the WireframeField class is compute() which
calculates the magnetic field and its spatial derivatives as a
user-provided set of points. The compute() method iterates through
the segments in the wireframe and calculated the contribution of
the current in each of those segments to each of the field points.
Inside the loop, a subroutine called wireframe_field_kernel() is
called to perform the Biot-Savart calculations for the each segment.
It therefore works in much the same way as the BiotSavart.compute()
method, although the latter iterates through Coil class instances
rather than segments and uses the biot_savart_kernal() internal
subfunction.