Implement dynamic wireframe constraint handling
Methods for dynamic constraint handling are implented in the
ToroidalWireframe class. Each constraint is stored in an OrderedDict
class instance. This allows the user to add and remove constraints
easily. Various methods are implemented to add and remove certain
types of constraints:

  - Continuity constraints: ensure current conservation at each node.
    These are added during initialization of the ToroidalWireframe
    class instance.

  - Poloidal current constraint: require the wireframe to have a
    certain net poloidal current in order to create a toroidal
    magnetic field.

  - Toroidal current constraint: require the wireframe to have a
    certain net toroidal current (creates a helical current
    distribution if combined with a poloidal current constraint)

  - Segment constraints: require certain segments to carry no current.
    This is useful, e.g. to reserve design space for ports.

In addition, a function enclosed_current(), is implented in the
WirefrmaeMagneticField class. This function numerically integrates the
projection of a magnet field onto the tangent vector of a closed curve
(specified as a CurveXYZFourier class instance) to determine the
enclosed current via Ampere's law. This is useful as a consistency
check to ensure that the toroidal and poloidal current constraints are
being upheld in wireframe solutions.