Add implicit segment constraints for wireframes
This commit addresses an issue that sometimes arose during constrained
least-squares solves on wireframe grids. The ToroidalWireframe
methods were previously programmed to mark a continuity constraint for
a node as redundant if and only if the node was connected to four
constrained segments.

If the node is connected to three constrained segments, the nominally
unconstrained segment must carry zero current by virtue of the
continuity constraint. The continuity constraint is not *necessarily*
redundant in this case... unless the nominally unconstrained segment
is connected at both ends to nodes with three explicitly constrained
segments. In this case, the nominally unconstrained ("trapped")
segment is constrained by two continuity constraints to carry zero
current, meaning that one of the continuity constraints is redundant.

To avoid this situation, trapped segments are now marked as
"implicitly constrained". Implicitly constrained segments are treated
the same as (explicitly) constrained segments when determining which
continuity constraints to mark as redundant. However, the
ToroidalWireframe class makes a distinction between explicitly and
implicitly constrained segments for accounting purposes. If the user
chooses to constrain a segment that happens to be already implicitly
constrained, it will become explicitly constrained.