Implement basic greedy coil optimizer
A new function, GSCO (Greedy Stellarator Coil Optimization) is
implemented in C++ in the new file wireframe_optimization.cpp. The
function adds loops of current to cells within the wireframe one by
one to reduce the squared flux objective.

To facilitate the optimization, the ToroidalWireframe Python class now
has methods to build and output a "cell key" which provides the
segments forming the loop around each cell in the wireframe.