Add method to visualize greedy wireframe optimizations
The method plot_cells_2d is added to the ToroidalWireframe class.
This represents the wireframe grid for a half period as a grid of
rectangular cells, color-coded according to how many current increments
have been added to the loop of segments around each cell. This
effectively shows the dipole moment contributed by each cell.