Introduce sparsity objective for GSCO
A new version of the Greedy Stellarator Coil Optimization (GCSO)
function is implemented that offers a second objective function in
addition to the objective that penalizes the normal magnetic field.
This new function, which is simply equal to the number of segments
in the grid with nonzero current (times 0.5), is meant to promote
sparsity. The user specifies a weighting factor, lambda_P, to
determine how much this objective is weighted relative to the normal
field objective.

In the present version, the function requires an array giving the
neighboring cells to each cell in the grid; however, this argument
is unnecessary and will be eliminated in a future version. It was
originally used to compute the new objective function, but the
definition of the new objective was subsequently simplified.

There is also a slight modification to how the `cells` input to GSCO is
defined. Previously, `cells` was assumed to only contain data for the
cells that were free to be optimized, i.e. none of the bordering
segments were constrained. In the present version, the array `cells`
is assumed to contain data for all cells, whether they are free or
not. There is an accompanying new argument, `free_cells`, which is a
logical array indicating which of the cells are free to be optimized.
This is meant to make the function more flexible, e.g. for multi-stage
optimizations.

In the ToroidalWireframe Python class, the method `get_cell_key` has
been modified to output all cell data as needed by the new version of
GSCO. There is also a new method, `get_free_cells`, which indicates
which cells are free (either via a list of indices or a logical array).
There is also a new method `get_cell_neighbors`, which returns an
array of the IDs of the neighbors of each cell.