Add area weighting to wireframe inductance matrix
The dBnormal_by_dsegmentcurrents_matrix function now provides the
option of weighting each matrix element by the square root of the
area of the plasma boundary ascribed to the corresponding quadrature
point. This way, the expression (A*x)^2, where A is the matrix and
x is the vector of segment currents, gives the integral of squared
flux over the plasma boundary.