Enable optimizations with no crossings
Solutions on the wireframe grid exhibiting the feature that all
currents form single-track loops (i.e. each node has either 2 or 0
active segments connected to it) are attractive for their simplicity.
This commit introduces some features that facilitate the attainment of
such solutions in both the greedy stellarator coil optimization (GSCO)
and regularized constrained least squares (RCLS) approaches.

In GSCO, the user may now opt to set a constraint that no current
crossings may appear in the grid (no_crossing=True). In this case,
loops of current will be considered ineligible to add if they create
current crossings, i.e. nodes with more than two active segments.
Eligibility checks, which now include the no-crossing check as well
as the max current check and the free loop check, have now been
factored out into a separate function. In addition, all eligibilty
checks are performed before starting the inner GSCO loop, such that
the inner loop is only run for eligible loops.

The GSCO function also has an option match_current, which allows the
optimizer to add loops to pre-existing coils with a current that
matches the current in the adjacent coil. However, this has not yet
been tested.

For the RCLS approach, the ToroidalWireframe class has new
functionality to enable optimizations of the currents in single-track
loops. In previous versions, if there were sufficient segment
constraints in the wireframe to create single-track loops, the
output constraint matrices would allow for no degrees of freedom.
This is because, for each single-track loop, one of the continuity
constraints is redundant. In the constraint_matrices function, the
user can now supply an optional input argument
assume_no_crossings=True, in which case one continuity constraint
per loop will be removed.