Add wrapper for the wireframe RCLS solver
A new function, `rcls_wireframe`, is added to
wireframe_optimization.py, designed to serve as a wrapper for the
`regularized_constrained_least_squares` function for ToroidalWireframe
class instances. Given an input wireframe, surface, and external field
(if applicable), calculates the inductance matrix, obtains constraint
matrices, and runs the low-level RCLS solver. To save time, the
function removes inactive segments from the system of equations be
input to the RCLS solver.

To support this function, the `constraint_matrices` method of the
ToroidalWireframe class has been revised to have the option of
removing equations for constrained segments.