Add wrapper for windowpane coil configurations
The function windowframe_wireframe() is added to wireframe.py. This
creates a ToroidalWireframe class instance and adds segment
constraints such that the free segments form a grid of "windowpane"
coils that are rectangular in the space of toroidal and poloidal
angles.

A new method free_segments() is added to the ToroidalWireframe class
that returns a list of the segments that are neither explicitly nor
implicitly constrained.

The constraint_matrices() method of the ToroialWireframe class is
modified such that it never outputs a list of free segments. Rather,
the user must call the free_segments() method separately.