Add maximum loop count option for GSCO
The GSCO function now has an input argument max_loop_count which, if
greater than zero, will set the maximum allowable loop_count value for
any loop in the wireframe.

This implemented as an optional argument in the Python wrapper
gsco_wireframe, and can be supplied as an optional parameter in
optimize_wireframe (both in wireframe_optimization.py).