Add option for no new coils in GSCO
The GSCO function now has the option to avoid adding new coils to a
wireframe during an optimization; i.e. for a loop in the wireframe to
be eligible for having its current incremented, at least one of its
segments must already be carrying current. The option may be set via
a new input argument when calling GSCO directly, or via an optional
parameter when calling through the optimize_wireframe Python wrapper.