Add abstract method mesh_representation to Port class
mesh_representation is added as an abstract method to the Port
abstract base class (ABC). Thus, any subclass (e.g. RectangularPort,
CircularPort) must contain an implementation of this method or else
an error will be raised. This is a necessary condition for the
functionality of the to_vtk method of the PortSet class, which calls
mesh_representation for each member Port class instance.