Reformulate wireframe constraint checker
The check_constraints method of the ToroidalWireframe class has been
modified to evaluate the constraints based on a linear combination of
an absolute tolerance and a relative tolerance. It is expected that the
introduction of a relative tolerance will reduce the likelihood that
the user would need to modify the default tolerance value for different
segment current magnitudes. The relative tolerance is multiplied by the
mean of the magnitudes of the segment currents. An absolute tolerance
is retained to support cases in which check_constraints is called when
the segment currents are all zero.