Improvements and fixes for wireframe optimizations
A number of improvements have been made to the wireframe optimization
procedures.

For the greedy stellarator coil optimization (GSCO) procedure:

  - The optimizer now handles the case in which no loops in the grid
    are eligible to have a loop of current added/subtracted without
    causing an error.

  - The optimizer detects when a solution has stopped improving by
    observing when a given iteration cancels the previous iteration.
    When this occurs, the optimization has reached a point where
    subsequent iterations would just oscillate.

  - An array 'iter_hist' is output to provide the iteration numbers
    of the entries of the other history arrays.

  - More GSCO parameters have been made optional and given default
    values.

  - GSCO output arrays are now cut down to the length of the actual
    number of iterations performed (i.e. prior to termination).

In addition, test_wf_optimization.py, a suite of unit tests for the
wireframe optimization procedures is added to tests/solve. Some of
the above improvements were motivated by bugs or other issues found
during the development of the tests.