Modify some variable names for consistency with paper
The names of some variables relevant to wireframe optimization have
been modified to be consistent with the equations in the paper on the
topic (K. C. Hammond, A framework for discrete optimization of
stellarator coils, https://arxiv.org/abs/2412.00267).

NOTE: One of these changes entails a modification to the API for the
wireframe optimization, so this commit is not backwards compatible.

The changes are as follows:

  - The regularization/weighting matrix for the RCLS objective function
    is now called "reg_W" in the dictionary of optimization parameters
    (this is the part that is not backwards compatible). Similar
    terminology is adopted within the internal RCLS functions.

  - The vector for the target normal field is now called "b" or
    "b_vec".

  - The matrix of constraint coefficients is now called "C" or
    "constraints_C".

Also, the docstring for wireframe_optimization is corrected to
accurately reflect the keys in the output dictionary.