Change GSCO history tracking procedure
The GSCO function and associated wrappers have been modified in order
to change how the history of an optimization is recorded. Previously,
the optimizer saved a copy of the intermediate solution and associated
objective function values from selected iterations. In the present
version, the function does not save copies of intermediate solutions,
but does save enough information to reconstruct the intermediate
solution from any iteration. To do this, the function saves two new
arrays: one that tracks to which wireframe cell a loop was current was
added in each iteration, and another that tracks how much current was
added in each iteration. Also, the objective function values are saved
for every iteration rather than only at certain intervals.

This change brings two advantages over the previous version. First,
the memory requirement should be lower in most cases, scaling as
n_iter rather than nSegments*n_history. Second, the saved data now
implicitly contain a complete record of every intermediate solution.

Alongside this modification, a new helper function,
`get_gsco_iteration` is added to reconstruct an intermediate solution
at the user's choice of iteration based on the records provided
by the GSCO function.

In addition, the interface to the GSCO function is slightly changed
to better reflect the new behavior. The parameter formerly known as
`nIter` is now `max_iter`, clarifying that it represents the maximum
number of iterations that might be run, rather than the number that
will actually be run. Further, the parameter formerly known as
`n_history` is now `print_interval`, since it no longer reflects how
much history data is stored but rather how frequently progress is
printed to the screen.

NOTE: due to the changes to the GSCO interface, this update is not
backward compatible.