Wrap scipy.linalg.qr function to catch bug
It has been observed that the scipy.linalg.qr function can return
erroneous results in some circumstances that appear to be dependent
on the platform, compiler, and/or software versions. In these cases,
it has been found that simply calling the function a second time can
return the correct output. To catch this bug (that may or may not
arise on a given machine or installation), a wrapper is written for
scipy.linalg.qr where it is used in the function
regularized_constrained_least_squares.

The cause of this issue is not well understood, but it has been
observed, for example, in the GitHub workflows for the simsopt
repository when the function is called after a magnetic field
evaluation has taken place.

For discussion on other, but possibly related issues:
    https://github.com/scipy/scipy/issues/5586
    https://github.com/numpy/numpy/issues/20356